+++ runners/s3-benchrunner-cpp/SdkClient.cpp
@@ -1,4 +1,4 @@
-#include "SdkClientRunner.h"
+#include "BenchmarkRunner.h"
 
 #include <fstream>
 #include <semaphore>
