+++ runners/s3-benchrunner-cpp/TransferManager.cpp
@@ -1,4 +1,4 @@
-#include "SdkTransferManagerRunner.h"
+#include "BenchmarkRunner.h"
 
 using namespace std;
